requirements.txt
python-dotenv

# huggingface
accelerate
torch
transformers

# rest api
flask==2.3.2
flask-cors==3.0.10
singleton-decorator==1.0.0